finished profile api an token

+++ src/app/api/users/profile/[id]/route.ts
@@ -1,8 +1,8 @@
 import { NextRequest, NextResponse } from "next/server";
 import prisma from "@/utils/db";
-
-import jwt from "jsonwebtoken";
-import { JWTPayload } from "@/utils/type";
+import { verifyToken } from "@/utils/verifyToken";
+import { UpdateUserDto } from "@/utils/dtos";
+import bcrypt from "bcryptjs";
 
 interface Props {
   params: { id: string };
@@ -15,7 +15,7 @@ interface Props {
  * @access private
  */
 
-export async function DELETE(requset: NextRequest, { params }: Props) {
+export async function DELETE(request: NextRequest, { params }: Props) {
   try {
     const user = await prisma.user.findUnique({
       where: { id: parseInt(params.id) },
@@ -25,23 +25,9 @@ export async function DELETE(requset: NextRequest, { params }: Props) {
       return NextResponse.json({ massage: "user Not found" }, { status: 404 });
     }
 
-    //  check user form token jwt
+    const userFromToken = verifyToken(request);
 
-    const authToken = requset.headers.get("authToken") as string;
-
-    if (!authToken) {
-      return NextResponse.json(
-        { massage: "not token provided ,acces denied" },
-        { status: 401 }
-      ); // 401 => unauthorized
-    }
-
-    const userFromToken = jwt.verify(
-      authToken,
-      process.env.JWT_SECRET as string
-    ) as JWTPayload;
-
-    if (userFromToken.id === user.id) {
+    if (userFromToken !== null && userFromToken.id === user.id) {
       await prisma.user.delete({ where: { id: parseInt(params.id) } });
       return NextResponse.json(
         { massage: "your profile (account) has been deleted" },
@@ -55,7 +41,99 @@ export async function DELETE(requset: NextRequest, { params }: Props) {
     ); // 403 => forbiddden
   } catch (error) {
     return NextResponse.json(
-      { massage: "internal server error" },
+      { massage: "internal server error from profile" },
+      { status: 500 }
+    );
+  }
+}
+
+/**
+ * @method GET
+ * @route http://localhost:3000/api/users/profile/:id
+ * @desc Get Profile By Id
+ * @access private
+ */
+
+export async function GET(request: NextRequest, { params }: Props) {
+  try {
+    const user = await prisma.user.findUnique({
+      where: { id: parseInt(params.id) },
+      select: {
+        id: true,
+        username: true,
+        email: true,
+        isAdmin: true,
+        createdAt: true,
+      },
+    });
+    if (!user) {
+      return NextResponse.json({ massage: "user Not found" }, { status: 404 });
+    }
+    const userFromToken = verifyToken(request);
+
+    if (userFromToken === null || userFromToken.id !== user.id) {
+      return NextResponse.json(
+        { massage: "you are not allowed, access denied" },
+        { status: 403 }
+      );
+    }
+
+    return NextResponse.json(user, { status: 200 });
+  } catch (error) {
+    return NextResponse.json(
+      { massage: "internal server error " },
+      { status: 500 }
+    );
+  }
+}
+
+/**
+ * @method PUT
+ * @route http://localhost:3000/api/users/profile/:id
+ * @desc Upate Profile By Id
+ * @access private
+ */
+
+export async function PUT(request: NextRequest, { params }: Props) {
+  try {
+    const user = await prisma.user.findUnique({
+      where: { id: parseInt(params.id) },
+    });
+    if (!user) {
+      return NextResponse.json({ massage: "user Not found" }, { status: 404 });
+    }
+
+    const userFromToken = verifyToken(request);
+
+    if (userFromToken === null || userFromToken.id !== user.id) {
+      return NextResponse.json(
+        { massage: "you are not allowed, access denied" },
+        { status: 403 }
+      );
+    }
+
+    const body = (await request.json()) as UpdateUserDto;
+
+    if (body.password) {
+      const salt = await bcrypt.genSalt(10);
+      body.password = await bcrypt.hash(body.password, salt);
+    }
+
+    const updatedUser = await prisma.user.update({
+      where: { id: parseInt(params.id) },
+      data: {
+        username: body.username,
+        email: body.email,
+        password: body.password,
+      },
+      select: {
+        password: false,
+      },
+    });
+    return NextResponse.json(updatedUser, { status: 200 });
+  } catch (error) {
+    return NextResponse.json(
+      { massage: "internal server error " },
       { status: 500 }
     );
   }
